API: Split up the high level devices into their own scripts.
